Modification to bash completion instructions
The previous instructions are tied to a specific version of bash-completion that is installed via brew. However it didn't specify in the instructions clearly that you needed to use brew to install bash-completion (When I read it, I thought I had to use brew to install bash).
The updated instructions clearly show that Homebrew is needed to first install bash-completion, and that the user should follow the instructions provided by the brew installer when updating the bash profile (since I suspect the exact command might change from version to version).